[#153] Move AWS dependencies file to AWS addon
