+++ src/commands/installAddon/index.ts
@@ -1,7 +1,7 @@
 import { Args, Flags, Command, ux } from '@oclif/core';
 
 import { postProcess } from '@/hooks/postProcess';
-import { requireAwsModules } from '@/templates/core/dependencies';
+import { requireAwsModules } from '@/templates/addons/aws/dependencies';
 import { awsModules } from '@/templates/core/types';
 
 import { GeneralOptions } from '../generate';
